+++ src/conv.jl
@@ -1,5 +1,3 @@
-# Define some new primitives: conv4 and pool
-
 """
 
     conv4(w, x; kwargs...)
@@ -146,12 +144,14 @@ end
 Unpooling; `reverse` of pooling.
 
 """
-function unpool{T}(x::KnetArray{T}; window=2)
+function unpool(x; window=2, o...)
+    if isa(window,Number); window=(window,window); end
     y = similar(x,updims(x; window=window))
-    Knet.poolx(y,x,x.*window^2; window=window,mode=1)
+    poolx(y,x,x.*prod(window); window=window, mode=1)
 end
 
-@primitive unpool(x;o...),dy,y -pool(-dy;o...)
+# @primitive unpool(x;o...),dy,y -pool(-dy;o...)
+@primitive  unpool(x;window=2,o...),dy,y  pool(dy;o...,window=window,mode=1)*prod(cdsize(window,ndims(x)-2))
 
 
 """
@@ -159,16 +159,16 @@ end
 Deconvolution; `reverse` of convolution.
 
 """
-function deconv4{T}(w::KnetArray{T},x::KnetArray{T}; o...)
+function deconv4(w,x; o...)
     y = similar(x,dcdims(w,x;o...))
     return conv4x(w,y,x;o...)
 end
 
-function deconv4w{T}(w::KnetArray{T},x::KnetArray{T},dy::KnetArray{T}; o...)
+function deconv4w(w,x,dy; o...)
     return conv4w(w,dy,x;o...)
 end
 
-function deconv4x{T}(w::KnetArray{T},x::KnetArray{T},dy::KnetArray{T}; o...)
+function deconv4x(w,x,dy; o...)
     return conv4(w,dy;o...)
 end
 
@@ -288,7 +288,8 @@ DT(::KnetArray{Float32})=UInt32(0)
 DT(::KnetArray{Float64})=UInt32(1)
 DT(::KnetArray{Float16})=UInt32(2)
 
-function cdims{T,N}(w::KnetArray{T,N},x::KnetArray{T,N}; padding=0, stride=1, o...)
+function cdims(w,x; padding=0, stride=1, o...)
+    N = ndims(x)
     ntuple(N) do i
         if i < N-1
             pi = (if isa(padding,Number); padding; else padding[i]; end)
@@ -302,7 +303,8 @@ function cdims{T,N}(w::KnetArray{T,N},x::KnetArray{T,N}; padding=0, stride=1, o.
     end
 end
 
-function pdims{T,N}(x::KnetArray{T,N}; window=2, padding=0, stride=window, o...)
+function pdims(x; window=2, padding=0, stride=window, o...)
+    N = ndims(x)
     ntuple(N) do i
         if i < N-1
             wi = (if isa(window,Number); window; else window[i]; end)
@@ -315,7 +317,8 @@ function pdims{T,N}(x::KnetArray{T,N}; window=2, padding=0, stride=window, o...)
     end
 end
 
-function dcdims{T,N}(w::KnetArray{T,N},x::KnetArray{T,N}; padding=0, stride=1, o...)
+function dcdims(w,x; padding=0, stride=1, o...)
+    N = ndims(x)
     ntuple(N) do i
         if i < N-1
             pi = (if isa(padding,Number); padding; else padding[i]; end)
@@ -329,12 +332,12 @@ function dcdims{T,N}(w::KnetArray{T,N},x::KnetArray{T,N}; padding=0, stride=1, o
     end
 end
 
-function updims{T,N}(x::KnetArray{T,N}; window=2)
-    if !isa(window,Number) error("Window size must be a number!") end
-    if !isa(window,Integer) error("Window size must be an integer!") end
+function updims(x; window=2)
+    if isa(window,Number); window=(window,window); end
+    N = ndims(x)
     ntuple(N) do i
         if i < N-1
-            size(x,i)*window
+            size(x,i)*window[i]
         else
             size(x,i)
         end
@@ -344,3 +347,260 @@ end
 # convolution padding size that preserves the input size when filter size is odd and stride=1
 padsize(w)=ntuple(i->div(size(w,i)-1,2), ndims(w)-2)
 
+
+### CPU convolution from Onur Kuru's CNN.jl
+
+function conv4{T}(w::Array{T,4}, x::Array{T,4};
+                  handle=nothing, alpha=1, beta=0,
+                  algo=0, workSpace=nothing, workSpaceSizeInBytes=0,
+                  padding=0, stride=1, upscale=1, mode=0)
+    # x: (W,H,C,N)
+    # w: (W,H,C,K) 
+    # y: (W,H,K,N) 
+    padding = isa(padding, Integer) ? [padding,padding] : collect(padding)
+    stride = isa(stride, Integer) ? [stride,stride] : collect(stride)
+    y = fill!(similar(x, cdims(w,x;padding=padding,stride=stride)),0)
+    Wx,Hx,Cx,N = size(x)
+    Ww,Hw,Cw,K = size(w)
+    if Cx!=Cw; throw(DimensionMismatch()); end
+    @inbounds for n in 1:N, k in 1:K, c in 1:Cx
+        y[:,:,k,n] += convy(view(x,:,:,c,n), view(w,:,:,c,k), padding, stride, mode)
+    end
+    if alpha != 1; y *= alpha; end
+    return y
+end
+
+function convy{T}(x0::SubArray{T,2}, w::SubArray{T,2}, padding::Array{Int,1}, stride::Array{Int,1}, mode)
+    x=x0
+    if any(padding .> 0) # this could be handled better....
+        x=zeros(eltype(x0), 2*padding+collect(size(x0))...)
+        x[padding[1]+1:end-padding[1],padding[2]+1:end-padding[2]] = x0
+    end
+    w1 = vec(w); if mode==0; w1 = reverse(w1); end
+    row_extend, col_extend = floor(Int, 1 + (collect(size(x)) - collect(size(w))) ./ stride)
+    widx = Int[sub2ind(size(x),i,j) for i in 1:stride[1]:size(x,1)-size(w,1)+1, j in 1:stride[2]:size(x,2)-size(w,2)+1] # linear indexes of filter positions in x
+    oidx = Int[sub2ind(size(x),i,j) for i in 1:size(w,1), j in 1:size(w,2)] # linear indexes of elements in a filter window
+    destidx = Int[i+(j-1) for i in vec(widx), j in vec(oidx)]
+    return reshape(x[destidx]*w1,row_extend,col_extend)
+end
+
+# dw = rot180(xcorr(x,dy))
+function conv4w{T}(w::Array{T,4}, x::Array{T,4}, dy::Array{T,4};
+                   handle=nothing, alpha=1, beta=0,
+                   algo=0, workSpace=nothing, workSpaceSizeInBytes=0,
+                   padding=0, stride=1, upscale=1, mode=0)
+    padding = isa(padding, Integer) ? [padding,padding] : collect(padding)
+    stride = isa(stride, Integer) ? [stride,stride] : collect(stride)
+    # x:    (Wx,Hx,Cx,N)
+    # dy:   (Wy,Hy,K,N) 
+    # dw:    (Ww,Hw,Cw,K) 
+    dw = fill!(similar(w),0)
+    Wx,Hx,C,Nx = size(x)
+    Wy,Hy,K,Ny = size(dy)
+    @inbounds for c in 1:C, k in 1:K, n in 1:Ny
+        dw[:,:,c,k] += convdw(view(x,:,:,c,n), view(dy,:,:,k,n), view(dw,:,:,c,k), padding, stride, mode)
+    end
+    if alpha != 1; dw *= alpha; end
+    return dw
+end
+
+# dw = rot180(xcorr(x,dy))
+function convdw{T}(x0::SubArray{T,2}, dy::SubArray{T,2}, w::SubArray{T,2}, padding::Array{Int,1}, stride::Array{Int,1}, mode::Int)
+    if any(padding .> 0) # this could be handled better...
+        x=zeros(eltype(x0), 2*padding+collect(size(x0))...)
+        x[padding[1]+1:end-padding[1],padding[2]+1:end-padding[2]] = x0
+    else
+        x=x0
+    end
+    x1l = last(collect(take(countfrom(1,stride[1]),size(dy,1))))
+    x2l = last(collect(take(countfrom(1,stride[2]),size(dy,2))))
+    widx = Int[sub2ind(size(x),i,j) for i in 1:size(w,1), j in 1:size(w,2)]
+    oidx = Int[sub2ind(size(x),i,j) for i in 1:stride[1]:x1l, j in 1:stride[2]:x2l] # linear indexes of elements in a filter window
+    destidx = Int[i+(j-1) for i in vec(widx), j in vec(oidx)]
+    y = reshape(x[destidx]*vec(dy),size(w))
+    if mode == 0; y = rot180(y); end
+    return y
+end
+
+# dx = xcorr(dy, w, 'full')
+function conv4x{T}(w::Array{T,4}, x::Array{T,4}, dy::Array{T,4};
+                   handle=nothing, alpha=1, beta=0,
+                   algo=0, workSpace=nothing, workSpaceSizeInBytes=0,
+                   padding=0, stride=1, upscale=1, mode=0)
+    padding = isa(padding, Integer) ? [padding,padding] : collect(padding)
+    stride = isa(stride, Integer) ? [stride,stride] : collect(stride)
+    dx = fill!(similar(x),0)
+    Wy,Hy,Ky,N = size(dy)
+    Ww,Hw,C,Kw = size(w)
+    if Ky!=Kw; throw(DimensionMismatch()); end
+    @inbounds for n in 1:N, c in 1:C, k in 1:Kw
+        dx[:,:,c,n] += convdx(view(dy,:,:,k,n), view(w,:,:,c,k), view(dx,:,:,c,n), padding, stride, mode)
+    end
+    if alpha != 1; dx *= alpha; end
+    return dx
+end
+
+# dx = xcorr(dy, w, 'full')
+function convdx{T}(dy::SubArray{T,2}, w::SubArray{T,2}, dx::SubArray{T,2}, padding::Array{Int,1}, stride::Array{Int,1}, mode::Int)
+    size_tdy = collect(size(dx)) + collect(size(w)) - 1 + 2padding
+    tdy = zeros(T, size_tdy...)
+    pad1, pad2 = map(x->x-1,size(w))
+    for (i,idy) in zip(countfrom(pad1+1,stride[1]), 1:size(dy,1)), (j,jdy) in zip(countfrom(pad2+1,stride[2]), 1:size(dy,2))
+        tdy[i,j] = dy[idy,jdy]
+    end
+    res = convy(view(tdy,:,:), view(w,:,:), [0,0], [1,1], 1-mode)
+    return all(padding .== 0) ? res : res[padding[1]+1:end-padding[1],padding[2]+1:end-padding[2]] 
+end
+
+
+#=
+function _conv2{T}(x::Array{T,2}, w::Array{T,2}; pad=0, stride=1, xcorr=false)
+    max_pad = map(x->x-1-pad,size(w))
+    y = conv2(x, xcorr ? rot180(w) : w)
+    return y[1+max_pad[1]:stride:end-max_pad[1], 1+max_pad[2]:stride:end-max_pad[2]]
+end
+=#
+
+#=
+function getConvolutionNdForwardOutputDim{T}(x::Array{T,4}, w::Array{T,4}; padding=padding,stride=stride)
+    padding = isa(padding, Integer) ? [padding,padding] : collect(padding)
+    stride = isa(stride, Integer) ? [stride,stride] : collect(stride)
+    Wx,Hx,Cx,N = size(x)
+    Ww,Hw,Cw,K = size(w)
+    @assert Cx==Cw
+    Wy,Hy = floor(Int, 1 + (Int[Wx,Hx] + 2*padding - Int[Ww,Hw]) ./ stride)
+    return (Wy,Hy,K,N)
+end
+=#
+
+
+### CPU pooling from Onur Kuru's CNN.jl
+
+
+function pool{T}(x::Array{T,4}; alpha=1, window=2, padding=0, stride=window, mode=0, o...)
+    # handle=nothing, beta=0, maxpoolingNanOpt=0,
+    stride = isa(stride, Integer) ? (stride, stride) : stride
+    window = isa(window, Integer) ? (window,window) : window
+    padding = isa(padding, Integer) ? (padding,padding) : padding
+    if any(map(x->x>0,padding))
+        x0=x
+        w,h,c,n = size(x0)
+        x=zeros(eltype(x0),w+2padding[1],h+2padding[2],c,n)
+        x[padding[1]+1:end-padding[1], padding[2]+1:end-padding[2],:,:] = x0
+    end
+    y = fill!(similar(x, pdims(x;window=window,padding=padding,stride=stride)), 0)
+    # x: (W,H,C,N)
+    Wx,Hx,C,Nx = size(x);
+    Wy,Hy,K,Ny = size(y);
+    if !(Nx == Ny && C==K); throw(DimensionMismatch()); end
+    if mode == 0
+        @inbounds for n in 1:Nx, c in 1:C, jy in 1:Hy, iy in 1:Wy
+            # iy, jy = div(i,stride[1])+1, div(j,stride[2])+1
+            i, j = 1+stride[1]*(iy-1), 1+stride[2]*(jy-1)
+            hx_end = j+window[2]-1 > Hx ? Hx : j+window[2]-1
+            wx_end = i+window[1]-1 > Wx ? Wx : i+window[1]-1
+            y[iy,jy,c,n] = maximum(x[i:wx_end,j:hx_end,c,n])
+        end
+    elseif mode == 1
+        @inbounds for n in 1:Nx, c in 1:C, jy in 1:Hy, iy in 1:Wy
+            # iy, jy = div(i,stride[1])+1, div(j,stride[2])+1
+            i, j = 1+stride[1]*(iy-1), 1+stride[2]*(jy-1)
+            hx_end = j+window[2]-1 > Hx ? Hx : j+window[2]-1
+            wx_end = i+window[1]-1 > Wx ? Wx : i+window[1]-1
+            y[iy,jy,c,n] = mean(x[i:wx_end,j:hx_end,c,n])
+        end
+    else
+        error("mode $mode not supported by cpu pool")
+    end
+    if alpha != 1; y *= alpha; end
+    return y
+end
+
+function poolx{T}(x::Array{T,4}, y::Array{T,4}, dy::Array{T,4}; alpha=1, beta=0,
+                  window=2, stride=window, padding=0, mode=0, # 0:max, 1:avg+pad, 2:avg-pad
+                  maxpoolingNanOpt=0, o...) # handle=cudnnhandle
+    stride = isa(stride, Integer) ? (stride, stride) : stride
+    window = isa(window, Integer) ? (window,window) : window
+    padding = isa(padding, Integer) ? (padding,padding) : padding
+    dx = fill!(similar(x), 0)
+    Wx,Hx,C,Nx = size(x);
+    Wy,Hy,K,Ny = size(y);
+    if mode == 0
+        @inbounds for n in 1:Nx, c in 1:C, i in 0:stride[1]:Wx-stride[1], j in 0:stride[2]:Hx-stride[2]
+            iy, jy = div(i,stride[1])+1, div(j,stride[2])+1
+            wx_end = min(i+window[1], Wx)
+            hx_end = min(j+window[2], Hx)
+            a = x[i+1:wx_end,j+1:hx_end,c,n]
+            di,dj = ind2sub(a,indmax(a))
+            dx[i+di,j+dj,c,n] += dy[iy,jy,c,n]
+        end
+    elseif mode == 1
+        @inbounds for n in 1:Nx, c in 1:C, i in 0:stride[1]:Wx-stride[1], j in 0:stride[2]:Hx-stride[2]
+            iy, jy = div(i,stride[1])+1, div(j,stride[2])+1
+            wx_end = min(i+window[1], Wx)
+            hx_end = min(j+window[2], Hx)
+            dx[i+1:wx_end,j+1:hx_end,c,n] += dy[iy,jy,c,n]
+        end
+        dx = dx ./ prod(window)
+    else
+        error("mode $mode not supported by cpu pool")
+    end
+    if alpha!=1; dx = alpha * dx; end
+    return dx
+end
+
+#=
+# mode == 0 maxpooling
+function poolx_buggy{T}(x::Array{T,4}, y::Array{T,4}, dy::Array{T,4};
+                  handle=nothing, alpha=1, beta=0, maxpoolingNanOpt=0,
+                  window=2, padding=0, stride=window, mode=0)
+    stride = isa(stride, Integer) ? (stride, stride) : stride
+    window = isa(window, Integer) ? (window,window) : window
+    padding = isa(padding, Integer) ? (padding,padding) : padding
+    dx = zeros(x)
+    if mode != 0; error("mode $mode not supported by cpu pool"); end
+    # x: (W,H,C,N)
+    if any(map(x->x>0,padding))
+        x0=x
+        w,h,c,n = size(x0)
+        x=zeros(eltype(x0),w+2padding[1],h+2padding[2],c,n)
+        x[padding[1]+1:end-padding[1], padding[2]+1:end-padding[2],:,:] = x0
+    end
+    dx1 = zeros(x)
+    Wx,Hx,C,Nx = size(x);
+    Wy,Hy,K,Ny = size(y);
+    if !(Nx == Ny && C==K); throw(DimensionMismatch()); end
+    # @inbounds for n in 1:Nx, c in 1:C, j in 1:stride[2]:Hx, i in 1:stride[1]:Wx
+    @inbounds for n in 1:Nx, c in 1:C, jy in 1:Hy, iy in 1:Wy
+        #= iy, jy = div(i,stride[1])+1, div(j,stride[2])+1
+        hx_end = j+window[2]-1 > Hx ? Hx : j+window[2]-1
+        wx_end = i+window[1]-1 > Hx ? Hx : i+window[1]-1 =#
+        i, j = 1+stride[1]*(iy-1), 1+stride[2]*(jy-1)
+        hx_end = j+window[2]-1 > Hx ? Hx : j+window[2]-1
+        wx_end = i+window[1]-1 > Wx ? Wx : i+window[1]-1
+        a = x[i:wx_end,j:hx_end,c,n]
+        di,dj = ind2sub(a,indmax(a))
+        # dx[i+di-1-padding[1],j+dj-1-padding[2],c,n] += dy[iy,jy,c,n]
+        dx1[i+di-1,j+dj-1,c,n] += dy[iy,jy,c,n]
+        any(map(x->x>0,padding)) && (dx[:,:,c,n] = dx1[padding[1]+1:end-padding[1],padding[2]+1:end-padding[2],c,n])
+    end
+    # @show dx1
+    # dx = dx1[padding[1]+1:end-padding[1],padding[2]+1:end-padding[2],:,:]
+    return dx
+end
+=#
+
+#=
+function getPoolingNdForwardOutputDim{T}(x::Array{T,4}; window=2, padding=0, stride=1, mode=0)
+    window = isa(window, Integer) ? (window,window) : window
+    padding = isa(padding, Integer) ? (padding,padding) : padding
+    stride = isa(stride, Integer) ? (stride,stride) : stride
+    @assert reduce(&, [w>p for (p,w) in zip(padding,window)])
+    dims = [size(x)...]
+    for i=1:length(dims)-2
+        # dims[i] = 1 + ceil((dims[i] + 2*padding[i] - window[i]) / stride[i])
+        dims[i] = length(1:stride[i]:dims[i]+padding[i])
+    end
+    tuple(dims...)
+end
+=#