+++ NEWS.md
@@ -13,6 +13,7 @@ General
 * Default conv4 padding changed back to 0. Knet.padsize(w) still available to compute input size preserving padding.
 * GPU garbage collection does not print '.', '+' unless user modifies gcinfo macro.
 * It is now an error to try to create a KnetArray when the active device is not a gpu.
+* CPU convolution and pooling operations implemented (@kuruonur1).
 
 Documentation and Testing
 -------------------------